# docs/swagger.yaml
basePath: /
definitions:
  core.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  core.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  core.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  core.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  core.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/core.PaginationResponse'
      status:
        type: string
      timestamp:
        type: string
    type: object
  core.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  core.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  core.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  core.UserRequest:
    properties:
      branch_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/core.UserRole'
    required:
    - email
    - full_name
    - password
    - phone
    - role
    type: object
  core.UserRole:
    enum:
    - admin_pusat
    - admin_cabang
    - kasir
    - teknisi
    - kurir
    - pelanggan
    - central_admin
    - branch_admin
    - cashier
    - technician
    - courier
    - customer
    type: string
    x-enum-varnames:
    - RoleAdminPusat
    - RoleAdminCabang
    - RoleKasir
    - RoleTeknisi
    - RoleKurir
    - RolePelanggan
host: localhost:8080
info:
  contact:
    email: support@iphoneservice.com
    name: iPhone Service API Support
    url: https://iphoneservice.com
  description: Backend API for iPhone Service Point of Sales system supporting 50
    branches across Indonesia
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: iPhone Service POS API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: List users with pagination and filters (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by branch ID
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users (admin)
      tags:
      - users
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user (admin)
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user (admin)
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user (admin)
      tags:
      - users
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change current user password
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiate password reset process
      parameters:
      - description: Forgot password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update current user profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using reset token
      parameters:
      - description: Reset password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"